+++ crates/uv-virtualenv/Cargo.toml
@@ -32,6 +32,7 @@ itertools = { workspace = true }
 pathdiff = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
+tempfile = "3.16.0"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 self-replace = { workspace = true }
